+++ src/Endpoint/Residential.php
@@ -8,7 +8,8 @@ use Aceproxies\ResellerApi\Exception\ApiException;
 use Aceproxies\ResellerApi\Exception\InvalidResponseException;
 use Aceproxies\ResellerApi\Exception\TransportException;
 use Aceproxies\ResellerApi\Http\HttpClientInterface;
-use Aceproxies\ResellerApi\Response\ResidentialCountriesResponse;
+use Aceproxies\ResellerApi\Response\Service\Residential\CountriesResponse;
+use Aceproxies\ResellerApi\Response\Service\Residential\RotationIntervalsResponse;
 
 final readonly class Residential implements ResidentialInterface
 {
@@ -23,12 +24,26 @@ final readonly class Residential implements ResidentialInterface
      * @throws InvalidResponseException
      * @throws TransportException
      */
-    public function countries(): ResidentialCountriesResponse
+    public function countries(): CountriesResponse
     {
         return $this->httpClient->request(
             HttpClientInterface::METHOD_GET,
             rtrim($this->baseUrl, '/') . '/api/v1/services/residential/countries',
-            ResidentialCountriesResponse::class,
+            CountriesResponse::class,
+        );
+    }
+
+    /**
+     * @throws ApiException
+     * @throws InvalidResponseException
+     * @throws TransportException
+     */
+    public function rotationIntervals(): RotationIntervalsResponse
+    {
+        return $this->httpClient->request(
+            HttpClientInterface::METHOD_GET,
+            rtrim($this->baseUrl, '/') . '/api/v1/services/residential/rotation-intervals/',
+            RotationIntervalsResponse::class,
         );
     }
 }